fixed circular dependency using application context
