--- src/main/java/cyber/login/jwt/system/loginsystemjwt/auth/AuthorizationService.java
@@ -1,23 +1,0 @@
-package cyber.login.jwt.system.loginsystemjwt.auth;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import cyber.login.jwt.system.loginsystemjwt.user.repositories.UserRepository;
-
-@Service
-public class AuthorizationService implements UserDetailsService{
-    @Autowired
-    UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        return userRepository.findByEmail(email);
-    } 
-
-
-    
-}
